Rename StoredEvent -> EloquentStoredEvent & StoredEventData -> StoredEvent
