+++ config/event-projector.php
@@ -46,7 +46,7 @@ return [
      * can change this to a class of your own. The only restriction is that
      * it should extend \Spatie\EventProjector\Models\StoredEvent.
      */
-    'stored_event_model' => \Spatie\EventProjector\Models\StoredEvent::class,
+    'stored_event_model' => \Spatie\EventProjector\Models\EloquentStoredEvent::class,
 
     /*
      * This class is responsible for storing events. To add extra behaviour you
