+++ src/EventHandlers/EventHandler.php
@@ -4,13 +4,13 @@ namespace Spatie\EventProjector\EventHandlers;
 
 use Exception;
 use Illuminate\Support\Collection;
-use Spatie\EventProjector\Models\StoredEventData;
+use Spatie\EventProjector\Models\StoredEvent;
 
 interface EventHandler
 {
     public function handles(): array;
 
-    public function handle(StoredEventData $event);
+    public function handle(StoredEvent $event);
 
     public function handleException(Exception $exception): void;
