+++ tests/TestClasses/AggregateRoots/AccountAggregateRootWithStoredEventSpecified.php
@@ -3,13 +3,13 @@
 namespace Spatie\EventProjector\Tests\TestClasses\AggregateRoots;
 
 use Spatie\EventProjector\AggregateRoot;
-use Spatie\EventProjector\Tests\TestClasses\Models\OtherStoredEvent;
+use Spatie\EventProjector\Tests\TestClasses\Models\OtherEloquentStoredEvent;
 use Spatie\EventProjector\Tests\TestClasses\AggregateRoots\StorableEvents\MoneyAdded;
 
 final class AccountAggregateRootWithStoredEventSpecified extends AggregateRoot
 {
     public $balance = 0;
-    public $storedEventModel = OtherStoredEvent::class;
+    public $storedEventModel = OtherEloquentStoredEvent::class;
 
     public function addMoney(int $amount): self
     {
