+++ tests/TestClasses/Models/OtherEloquentStoredEvent.php
@@ -2,9 +2,9 @@
 
 namespace Spatie\EventProjector\Tests\TestClasses\Models;
 
-use Spatie\EventProjector\Models\StoredEvent as BaseStoredEvent;
+use Spatie\EventProjector\Models\EloquentStoredEvent as BaseStoredEvent;
 
-class OtherStoredEvent extends BaseStoredEvent
+class OtherEloquentStoredEvent extends BaseStoredEvent
 {
     /**
      * The table associated with the model.
