+++ tests/TestClasses/Projectors/ProjectorThatWritesMetaData.php
@@ -4,7 +4,7 @@ namespace Spatie\EventProjector\Tests\TestClasses\Projectors;
 
 use Spatie\EventProjector\Projectors\Projector;
 use Spatie\EventProjector\StoredEventRepository;
-use Spatie\EventProjector\Models\StoredEventData;
+use Spatie\EventProjector\Models\StoredEvent;
 use Spatie\EventProjector\Projectors\ProjectsEvents;
 use Spatie\EventProjector\Tests\TestClasses\Events\MoneyAddedEvent;
 
@@ -18,7 +18,7 @@ final class ProjectorThatWritesMetaData implements Projector
 
     protected $trackStream = '*';
 
-    public function onMoneyAdded(StoredEventData $storedEventData, MoneyAddedEvent $event)
+    public function onMoneyAdded(StoredEvent $storedEventData, MoneyAddedEvent $event)
     {
         $storedEventData->meta_data['user_id'] = 1;
